Add Frontend

+++ database/migrations/2020_01_31_154629_create_stall_equipment_table.php
@@ -16,7 +16,7 @@ class CreateStallEquipmentTable extends Migration
         Schema::create('stall_equipment', function (Blueprint $table) {
             $table->increments('id');
             $table->json('name');
-            $table->float('price');
+            $table->unsignedInteger('price');
         });
     }
